# Conflicts:
#	The Cook of Village/Assets/Prefabs/UI/Restaurant/RestaurantCanvas.prefab
#	The Cook of Village/Assets/Scenes/Restaurant.unity

+++ The Cook of Village/Assets/Scripts/UI/Naeun/Shop/ShopSelect.cs
@@ -24,7 +24,7 @@ public class ShopSelect : MonoBehaviour
             CountSlider.value = 0;
             CountSlider.maxValue = ItemData.Instance.MaxMaterialCount;
             BuyMaxCount = ItemData.Instance.MaxMaterialCount - currentCount;
-            MoneyMaxCount = GameData.Instance.Money / slot.Infos.Price;
+            MoneyMaxCount = (int)GameData.Instance.Money / slot.Infos.Price;
         }
     }
     private int BuyMaxCount;
